requirements.txt
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
jinja2>=3.1.2
colorama>=0.4.6
pydantic>=2.0.0
